Rename bokeh.crossfilter.{objects->models}
