+++ bokeh/crossfilter/models.py
