+++ examples/app/crossfilter/crossfilter_app.py
@@ -2,7 +2,7 @@ from bokeh.plot_object import PlotObject
 from bokeh.server.utils.plugins import object_page
 from bokeh.server.app import bokeh_app
 from bokeh.plotting import curdoc, cursession
-from bokeh.crossfilter.objects import CrossFilter
+from bokeh.crossfilter.models import CrossFilter
 from bokeh.sampledata.autompg import autompg
 
 @bokeh_app.route("/bokeh/crossfilter/")
